requirements.txt
Django==1.11.5
PyJWT==1.5.3
django-braces==1.11.0
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
oauthlib==1.1.2
pytz==2017.2
six==1.10.0
wsgiref==0.1.2
